feat(tms): 实现头程申请单合并功能
- 新增合并头程申请单的接口和实现逻辑
- 添加合并申请单时的校验逻辑，包括申请单是否存在、状态是否允许合并等
- 实现将多个申请单合并为一个新的头程单的功能- 优化了错误码和异常处理

+++ yudao-module-tms/yudao-module-tms-biz/src/main/java/cn/iocoder/yudao/module/tms/controller/admin/first/mile/item/vo/TmsFirstMileItemSaveReqVO.java
@@ -17,10 +17,10 @@ public class TmsFirstMileItemSaveReqVO {
     private Long id;
 
     @Schema(description = "申请项ID")
-    private Integer requestItemId;
+    private Long requestItemId;
 
     @Schema(description = "产品ID")
-    private Integer productId;
+    private Long productId;
 
     @Schema(description = "箱数")
     private Integer boxQty;
